Move NonPredictiveRecommender to new subpackage
