+++ baybe/recommenders/nonpredictive/__init__.py
@@ -1,0 +1,1 @@
+"""Nonpredictive recommenders."""
