+++ tests/serialization/test_naive_hybrid_serialization.py
@@ -3,12 +3,12 @@
 import pytest
 
 from baybe.campaign import Campaign
-from baybe.recommenders.base import NonPredictiveRecommender
 from baybe.recommenders.bayesian import (
     BayesianRecommender,
     NaiveHybridRecommender,
     SequentialGreedyRecommender,
 )
+from baybe.recommenders.nonpredictive.base import NonPredictiveRecommender
 from baybe.searchspace import SearchSpaceType
 from baybe.utils.basic import get_subclasses
